track the running auction inside auctioneer
tracking and delegating to directly to the live
auction inside auctioneer simplifies the
auction manger significantly, turning it into
just an auction factory.
